+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - New LabbcatEdit functions:
   + *uploadParticipantAttributes* - upload participant attribute data from CSV file
+  + *uploadTranscriptAttributes* - upload transcript attribute data from CSV file
 
 Requires LaBB-CAT version 20260812.2001